first version of seller inventory + little change on login
