+++ views/inventory-edit.hbs
@@ -1,0 +1,70 @@
+<ul class="nav nav-tabs">
+    <li class="active"><a data-toggle="tab" href="#home">Home</a></li>
+    <li><a data-toggle="tab" href="#menu1">Menu 1</a></li>
+    <li><a data-toggle="tab" href="#menu2">Menu 2</a></li>
+</ul>
+
+<div class="tab-content">
+    <div id="home" class="tab-pane fade in active">
+        <h3></h3>
+        <p>Update your inventory product</p>
+        {{!-- **************** --}}
+        <div class='purchaseContainer'>
+            <div class='productImg'>
+                {{!-- <img src='/images/simpsons.png' /> --}}
+                <img src='{{product.itemImgPath}}' />
+            </div>
+            <div class='productInfoBox'>
+                <div class='productInfo'>
+                    <header class='purchaseHeader'>{{product.itemName}}
+                        <h6 class='purchaseH6'>by {{product.itemSeller}}</h6>
+                    </header>
+                    <hr />
+                    <div>
+                        <h6>Price:
+                            <span style='font-size:20px; color:#D63D28;'>$ {{product.itemPrice}}</span>
+                        </h6>
+                    </div>
+                    <div>
+                        <h4 class='purchaseH4' style='color:green;'>In Stock</h4>
+
+                    </div>
+                    <div>
+                        <h4 class='purchaseH4'>About the product</h4>
+                        <ul class='purList' style="list-style-type:disc">
+                            <li>{{product.itemDescription}}</li>
+                            <li>Test 2</li>
+                        </ul>
+                    </div>
+                </div>
+            </div>
+            <div class='purchaseSection'>
+                <div class='purchaseBox'>
+                    <div class='purBoxFit'>
+                        <span class='purchaseQty'>Qty
+                            <select id='purSelect'>
+                                <option value='1'>1</option>
+                                <option value='2'>2</option>
+                                <option value='3'>3</option>
+                                <option value='4'>4</option>
+                                <option value='5'>5</option>
+                                <option value='6'>6</option>
+                            </select>
+                            <a href=''>cart:</a><span id='cartQty'>0</span>
+                            {{!-- <button id='cartBtn' type="button">cart:</button> <span id='cartQty'>0</span> --}}
+                        </span>
+                        <button id='incQtyBtn' type='button' data-id='{{this._id}}'>Add to Cart</button>
+                    </div>
+                </div>
+            </div>
+            <script src='javascripts/purchase.js'></script>
+        </div>
+
+
+        {{!-- *********** --}}
+    </div>
+    <div id="menu1" class="tab-pane fade">
+        <h3>New Product</h3>
+        <p>Add new product to inventory</p>
+    </div>
+</div>
