+++ views/inventory.hbs
@@ -15,11 +15,11 @@
                 </ul>
             </div>
             <div class="col-sm-4 col-md-3 col-lg-2 col-xl-2" style="border:1px solid white">
-                <button id="buyBtn" class="btn btn-outline-primary ml-5" type="button">learn more</button>
+                <form action="/inventories/update/{{this._id}}" method="POST">
+                    <button type="submit" id="buyBtn" class="btn btn-outline-primary ml-5" type="button">edit</button>
+                </form>
             </div>
         </div>
-
-
         <hr />
     </div>
     {{/each}}
