+++ views/login.hbs
@@ -39,6 +39,12 @@
 	{{msg}}
 </div>
 
-{{!--
-<!--can add javascript fle here or end of doc-->
-<script src="javascripts/login-register.js"></script> --}}
+{{#if msg}}
+<div class=" container-fluid alert alert-danger col-lg-4  ml-50">
+	<p><strong>Erorr: </strong>{{msg}}</p>
+</div>
+{{/if}} {{#if msgok}}
+<div class=" container-fluid alert alert-success col-lg-4  ml-50">
+	<p><strong>congratulation: </strong>{{msgok}}</p>
+</div>
+{{/if}}
